Cleanup deletion recording utils
